Refactor dictionary runtime and breakpoint processor
Refactor breakpoint processor iterative path for maintainability #33
Refactor dictionary runtime into focused modules and shared detector helpers #44
